Add user's desired amount of squares to the grid
createDivs(currentPromptValue) creates the user's desired amount of
squares per row/column, it does this by setting each square's width to
the exact amount of space needed for the container to wrap in each row,
and also set's the square's height for the exact amount to be filled in
the container. It is run at the beginning of the page load with a
default amount of 16 grids.

At the start of the button prompt, a for loop deletes the current
square's in the canvas, in order to not overflow the canvas with the
new squares being added.

Each new square's border overflowed with the one next to it, so the
border width was set to 0.5px in order to have a total width of 1px, and
added a border to container with a 0.5px width so that the squares with
no other adjacent squares has an equal border to the rest. Added 1px
border to button for symmetry.