Add hover effect on click
Adds three event listeners to the dynamicDiv when they're created, and
a hover variable, this is going to apply the hover effect only when the
user is clicking down on their mouse.

Also adds an ondragstart and ondrop attribute to the body element, and
returns false on either event, this is so that when the user is holding
click it doesn't start dragging an element, disrupting the hover event.